+++ examples/replay/replay_klws_agent.py
@@ -33,13 +33,9 @@ def copytree(src, dst, symlinks=False, ignore=None):
             if os.path.lexists(d):
                 os.remove(d)
             os.symlink(os.readlink(s), d)
-            try:
-                st = os.lstat(s)
-                mode = stat.S_IMODE(st.st_mode)
-                os.lchmod(d, mode)
-            except Exception as e:
-                print(e)
-                pass  # lchmod not available
+            st = os.lstat(s)
+            mode = stat.S_IMODE(st.st_mode)
+            os.chmod(d, mode, follow_symlinks=False)
         elif os.path.isdir(s):
             copytree(s, d, symlinks, ignore)
         else:
@@ -81,7 +77,6 @@ def main(scenarios, sim_name, headless, seed, speed, max_steps, save_dir, write)
         agent_specs={AGENT_ID: agent_spec},
         sim_name=sim_name,
         headless=headless,
-        visdom=False,
         timestep_sec=0.1,
         sumo_headless=True,
         seed=seed,